Rename State => Store
