+++ code/state/src/types.ts
@@ -1,19 +1,19 @@
 /**
  * Basic shape of an event fired through the state-machine.
  */
-export type IStateEvent = {
+export type IStoreEvent = {
   type: string;
   payload: object;
 };
 
 /**
- * A wrapper of an `IStateEvent` that is dispatched through
- * the [events$] observable.
+ * A wrapper of an [IStoreEvent] that is dispatched through
+ * the store's `events$` observable.
  */
 export type IDispatch<
   M extends {} = {},
-  E extends IStateEvent = IStateEvent,
-  SE extends IStateEvent = IStateEvent
+  E extends IStoreEvent = IStoreEvent,
+  SE extends IStoreEvent = IStoreEvent
 > = {
   type: E['type'];
   payload: E['payload'];
@@ -23,9 +23,9 @@ export type IDispatch<
 };
 
 /**
- * A notification of a change to the state tree.
+ * A notification of a change to a store's state-tree.
  */
-export type IStateChange<M extends {} = {}, E extends IStateEvent = IStateEvent> = {
+export type IStateChange<M extends {} = {}, E extends IStoreEvent = IStoreEvent> = {
   event: E;
   from: M;
   to: M;
